Split AHF tests into separate files.
