+++ tests/frontends/test_ahf_3.py
@@ -1,21 +1,8 @@
 from ytree.frontends.ahf import \
-    AHFArbor, \
-    AHFCRMArbor
+    AHFArbor
 from ytree.utilities.testing import \
     ArborTest, \
     TempDirTest
-
-class AHFArborTest(TempDirTest, ArborTest):
-    arbor_type = AHFArbor
-    test_filename = "ahf_halos/snap_N64L16_000.parameter"
-    num_data_files = 136
-    tree_skip = 100
-
-class AHFCRMArborTest(TempDirTest, ArborTest):
-    arbor_type = AHFCRMArbor
-    test_filename = "AHF_100_tiny/GIZMO-NewMDCLUSTER_0047.snap_128.parameter"
-    num_data_files = 5
-    tree_skip = 100
 
 class AHFNameConfigArborTest(TempDirTest, ArborTest):
     arbor_type = AHFArbor
